+++ app/Auth/strategies.js
@@ -1,41 +1,12 @@
 /**
- * Authstrategies
+ * Auth strategies
  */
 const passport = require('passport')
 const GithubStrategy = require('passport-github').Strategy
 const FacebookStrategy = require('passport-facebook').Strategy
 
-const {
-  users,
-  strategies
-} = require('config/auth')
-
-const transformProfile = profile => {
-  const [{ value }] = profile.emails
-
-  return {
-    email: value,
-    name: profile.displayName || profile.name.givenName + ' ' + profile.name.familyName
-  }
-}
-
-const handleProfile = (token, refreshToken, profile, done) => {
-  const { email, name } = transformProfile(profile)
-  const { model } = users
-
-  model.findOne({email})
-    .then(user => user ? user.toJSON() : null)
-    .then(user => {
-      if (user) {
-        return done(null, user)
-      }
-
-      return model.create(name, email)
-        .then(user => done(null, user))
-        .catch(error => console.log('error on create:', error))
-    })
-    .catch(error => console.log(error))
-}
+const handleProfile = require('./register')
+const { strategies } = require('config/auth')
 
 const handleRoutesStrategies = (strategy, Router, controller) => {
   Router.get(`/${strategy}`, passport.authenticate(strategy))